Multi agent results near previous quality
I’m getting multi agent results that are near the quality of my single
agent setup.  Add some tests where both players are random.  Where the
starting player is chosen randomly, they should be 50/50.  Where the
starting player is always the same, that player should have a 58% win
rate.  This is the case here, which is very good.